Adds check missing granules functionality
Adds functionality to check for and download missing granules.

This change introduces command-line arguments to check for missing
granules in the download directory based on the payload file.
It can also automatically download the missing granules if the
`--download-missing` flag is provided.
Missing granules are determined by comparing the expected files
from the payload with the actual files present in the download
directory. A JSON file containing the list of missing granules
is also generated.